Refactor authentication and book management components
- Moved Login component to a new directory structure under features/auth.
- Created a new AuthForm component to handle authentication forms for login and registration.
- Introduced ModeToggle component for switching between login and registration modes.
- Implemented useAuth hook for managing authentication state and API interactions.
- Created Books component with BookForm and BookTable for managing book records.
- Added useBooks hook for handling book-related API calls and state management.
- Removed legacy components and organized the codebase for better maintainability.
