--- client/src/components/Login.js
@@ -1,283 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { authAPI } from "../services/api";
-
-const Login = ({ onLogin, initialError, onClearError }) => {
-  const [mode, setMode] = useState("login"); // "login" or "register"
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-    passwordConfirmation: "",
-    name: "",
-    userType: "member",
-  });
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false);
-
-  // Set initial error when component mounts or when initialError changes
-  useEffect(() => {
-    if (initialError) {
-      setError(initialError);
-    }
-  }, [initialError]);
-
-  const handleModeChange = (newMode) => {
-    setMode(newMode);
-    setError("");
-    if (onClearError) {
-      onClearError();
-    }
-    setFormData({
-      email: "",
-      password: "",
-      passwordConfirmation: "",
-      name: "",
-      userType: "member",
-    });
-  };
-
-  const handleChange = (e) => {
-    // Clear error when user starts typing
-    if (error && onClearError) {
-      onClearError();
-    }
-    setError("");
-
-    setFormData({
-      ...formData,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const validateForm = () => {
-    if (mode === "register") {
-      if (!formData.name.trim()) {
-        setError("Name is required");
-        return false;
-      }
-      if (formData.password !== formData.passwordConfirmation) {
-        setError("Passwords do not match");
-        return false;
-      }
-      if (formData.password.length < 6) {
-        setError("Password must be at least 6 characters long");
-        return false;
-      }
-    }
-    return true;
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) {
-      return;
-    }
-
-    setLoading(true);
-    setError("");
-
-    try {
-      let response;
-
-      if (mode === "login") {
-        response = await authAPI.login(formData.email, formData.password);
-      } else {
-        // Register mode
-        const userData = {
-          email: formData.email,
-          password: formData.password,
-          password_confirmation: formData.passwordConfirmation,
-          name: formData.name,
-          user_type: formData.userType,
-        };
-        response = await authAPI.register(userData);
-      }
-
-      if (response.data.success) {
-        // Store user data and JWT token in localStorage
-        localStorage.setItem("user", JSON.stringify(response.data.user));
-        localStorage.setItem("authToken", response.data.token);
-        onLogin(response.data.user);
-      }
-    } catch (err) {
-      if (mode === "register" && err.response?.data?.errors) {
-        setError(err.response.data.errors.join(", "));
-      } else {
-        setError(
-          err.response?.data?.message ||
-            `${mode === "login" ? "Login" : "Registration"} failed`
-        );
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            Library Management System
-          </h2>
-        </div>
-
-        {/* Mode Toggle */}
-        <div className="flex justify-center">
-          <div className="bg-gray-200 rounded-lg p-1">
-            <button
-              onClick={() => handleModeChange("login")}
-              className={`px-4 py-2 text-sm font-medium rounded-md transition-colors duration-200 ${
-                mode === "login"
-                  ? "bg-white text-gray-900 shadow-sm"
-                  : "text-gray-600 hover:text-gray-900"
-              }`}
-            >
-              Sign In
-            </button>
-            <button
-              onClick={() => handleModeChange("register")}
-              className={`px-4 py-2 text-sm font-medium rounded-md transition-colors duration-200 ${
-                mode === "register"
-                  ? "bg-white text-gray-900 shadow-sm"
-                  : "text-gray-600 hover:text-gray-900"
-              }`}
-            >
-              Register
-            </button>
-          </div>
-        </div>
-
-        <div className="bg-white py-8 px-6 shadow rounded-lg">
-          {error && (
-            <div className="mb-4 bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded-md">
-              {error}
-            </div>
-          )}
-
-          <form className="space-y-6" onSubmit={handleSubmit}>
-            {mode === "register" && (
-              <div>
-                <label
-                  htmlFor="name"
-                  className="block text-sm font-medium text-gray-700"
-                >
-                  Full Name
-                </label>
-                <input
-                  id="name"
-                  name="name"
-                  type="text"
-                  required
-                  className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                  value={formData.name}
-                  onChange={handleChange}
-                />
-              </div>
-            )}
-
-            <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium text-gray-700"
-              >
-                Email Address
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                value={formData.email}
-                onChange={handleChange}
-              />
-            </div>
-
-            <div>
-              <label
-                htmlFor="password"
-                className="block text-sm font-medium text-gray-700"
-              >
-                Password
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete={
-                  mode === "login" ? "current-password" : "new-password"
-                }
-                required
-                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                value={formData.password}
-                onChange={handleChange}
-              />
-            </div>
-
-            {mode === "register" && (
-              <>
-                <div>
-                  <label
-                    htmlFor="passwordConfirmation"
-                    className="block text-sm font-medium text-gray-700"
-                  >
-                    Confirm Password
-                  </label>
-                  <input
-                    id="passwordConfirmation"
-                    name="passwordConfirmation"
-                    type="password"
-                    autoComplete="new-password"
-                    required
-                    className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                    value={formData.passwordConfirmation}
-                    onChange={handleChange}
-                  />
-                </div>
-
-                <div>
-                  <label
-                    htmlFor="userType"
-                    className="block text-sm font-medium text-gray-700"
-                  >
-                    User Type
-                  </label>
-                  <select
-                    id="userType"
-                    name="userType"
-                    className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                    value={formData.userType}
-                    onChange={handleChange}
-                  >
-                    <option value="member">Member</option>
-                    <option value="librarian">Librarian</option>
-                  </select>
-                </div>
-              </>
-            )}
-
-            <div>
-              <button
-                type="submit"
-                disabled={loading}
-                className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                {loading
-                  ? mode === "login"
-                    ? "Signing In..."
-                    : "Creating Account..."
-                  : mode === "login"
-                  ? "Sign In"
-                  : "Sign Up"}
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Login;
