+++ client/src/features/auth/components/AuthForm.js
@@ -1,0 +1,207 @@
+import React, { useState, useEffect } from "react";
+
+const AuthForm = ({
+  mode,
+  onSubmit,
+  onCancel,
+  loading,
+  error,
+  initialError,
+  onClearError,
+}) => {
+  const [formData, setFormData] = useState({
+    email: "",
+    password: "",
+    passwordConfirmation: "",
+    name: "",
+    userType: "member",
+  });
+
+  // Set initial error when component mounts or when initialError changes
+  useEffect(() => {
+    if (initialError && onClearError) {
+      onClearError();
+    }
+  }, [initialError, onClearError]);
+
+  const handleChange = (e) => {
+    // Clear error when user starts typing
+    if (error && onClearError) {
+      onClearError();
+    }
+
+    setFormData({
+      ...formData,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const validateForm = () => {
+    if (mode === "register") {
+      if (!formData.name.trim()) {
+        return "Name is required";
+      }
+      if (formData.password !== formData.passwordConfirmation) {
+        return "Passwords do not match";
+      }
+      if (formData.password.length < 6) {
+        return "Password must be at least 6 characters long";
+      }
+    }
+    return null;
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+
+    const validationError = validateForm();
+    if (validationError) {
+      if (onClearError) {
+        onClearError();
+      }
+      return;
+    }
+
+    const userData = {
+      email: formData.email,
+      password: formData.password,
+      password_confirmation: formData.passwordConfirmation,
+      name: formData.name,
+      user_type: formData.userType,
+    };
+
+    await onSubmit(userData);
+  };
+
+  return (
+    <div className="bg-white py-8 px-6 shadow rounded-lg">
+      {error && (
+        <div className="mb-4 bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded-md">
+          {error}
+        </div>
+      )}
+
+      <form className="space-y-6" onSubmit={handleSubmit}>
+        {mode === "register" && (
+          <div>
+            <label
+              htmlFor="name"
+              className="block text-sm font-medium text-gray-700"
+            >
+              Full Name
+            </label>
+            <input
+              id="name"
+              name="name"
+              type="text"
+              required
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+              value={formData.name}
+              onChange={handleChange}
+            />
+          </div>
+        )}
+
+        <div>
+          <label
+            htmlFor="email"
+            className="block text-sm font-medium text-gray-700"
+          >
+            Email Address
+          </label>
+          <input
+            id="email"
+            name="email"
+            type="email"
+            autoComplete="email"
+            required
+            className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+            value={formData.email}
+            onChange={handleChange}
+          />
+        </div>
+
+        <div>
+          <label
+            htmlFor="password"
+            className="block text-sm font-medium text-gray-700"
+          >
+            Password
+          </label>
+          <input
+            id="password"
+            name="password"
+            type="password"
+            autoComplete={
+              mode === "login" ? "current-password" : "new-password"
+            }
+            required
+            className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+            value={formData.password}
+            onChange={handleChange}
+          />
+        </div>
+
+        {mode === "register" && (
+          <>
+            <div>
+              <label
+                htmlFor="passwordConfirmation"
+                className="block text-sm font-medium text-gray-700"
+              >
+                Confirm Password
+              </label>
+              <input
+                id="passwordConfirmation"
+                name="passwordConfirmation"
+                type="password"
+                autoComplete="new-password"
+                required
+                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+                value={formData.passwordConfirmation}
+                onChange={handleChange}
+              />
+            </div>
+
+            <div>
+              <label
+                htmlFor="userType"
+                className="block text-sm font-medium text-gray-700"
+              >
+                User Type
+              </label>
+              <select
+                id="userType"
+                name="userType"
+                className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+                value={formData.userType}
+                onChange={handleChange}
+              >
+                <option value="member">Member</option>
+                <option value="librarian">Librarian</option>
+              </select>
+            </div>
+          </>
+        )}
+
+        <div>
+          <button
+            type="submit"
+            disabled={loading}
+            className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            {loading
+              ? mode === "login"
+                ? "Signing In..."
+                : "Creating Account..."
+              : mode === "login"
+              ? "Sign In"
+              : "Sign Up"}
+          </button>
+        </div>
+      </form>
+    </div>
+  );
+};
+
+export default AuthForm;
